refactor: move computation modules to services/compute/ and rename lookups
- Move indicators, portfolio, pseudo_etf, watchlist into services/compute/
- Rename lookups.py to entity_lookups.py for clarity
- Update all imports across routers, services, and tests
