+++ backend/app/main.py
@@ -15,7 +15,7 @@ from app.database import async_session, engine, Base
 from app.models import Asset  # noqa: F401 - ensure models are imported for create_all
 from app.routers import annotations, assets, groups, holdings, portfolio, prices, pseudo_etfs, pseudo_etf_analysis, quotes, search, settings as settings_router, tags, thesis, watchlist
 from app.services.price_sync import sync_all_prices
-from app.services.watchlist import compute_and_cache_indicators
+from app.services.compute.watchlist import compute_and_cache_indicators
 from app.services.yahoo import batch_fetch_currencies
 
 logger = logging.getLogger(__name__)
