+++ backend/app/routers/annotations.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.database import get_db
-from app.services.lookups import get_asset
+from app.services.entity_lookups import get_asset
 from app.schemas.annotation import AnnotationCreate, AnnotationResponse
 from app.services import annotation_service
