+++ backend/app/services/compute/indicators.py
